refactor: вынесен базовый класс кастомного текстового инпута
