+++ src/app/shared/components/inputs/password-input/format-password-requirement.pipe.ts
