+++ src/app/shared/components/inputs/text-input-base.ts
@@ -1,39 +1,25 @@
-import { NgIf } from '@angular/common';
 import {
-  ChangeDetectionStrategy,
-  Component,
+  Directive,
   DoCheck,
   ElementRef,
-  HostListener,
   Input,
   ViewChild,
 } from '@angular/core';
 import { ControlValueAccessor, NgControl } from '@angular/forms';
-import { ButtonModule } from 'primeng/button';
-import { InputTextModule } from 'primeng/inputtext';
-import { OverlayPanel, OverlayPanelModule } from 'primeng/overlaypanel';
+import { OverlayPanel } from 'primeng/overlaypanel';
 
-@Component({
-  selector: 'app-text-input',
-  standalone: true,
-  imports: [InputTextModule, OverlayPanelModule, ButtonModule, NgIf],
-  templateUrl: './text-input.component.html',
-  styleUrl: './text-input.component.scss',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class TextInputComponent implements DoCheck, ControlValueAccessor {
+@Directive()
+export abstract class TextInputComponentBase
+  implements DoCheck, ControlValueAccessor
+{
   @Input() public withClearButton: boolean = true;
 
   @Input() public placeholder: string = '';
-
-  public disabled: boolean = false;
 
   public get errorMessage(): string {
     switch (true) {
       case this.control.hasError('required'):
         return 'Поле обязательно для заполнения';
-      case this.control.hasError('email'):
-        return 'Некорректный формат почты';
       case this.control.hasError('minlength'):
         return `Минимальная длина ${this.control.getError('minlength')?.requiredLength} символов`;
       case this.control.hasError('maxlength'):
@@ -47,15 +33,17 @@ export class TextInputComponent implements DoCheck, ControlValueAccessor {
     return this.value;
   }
 
-  @ViewChild('errorPanel') private errorPanel!: OverlayPanel;
+  public disabled: boolean = false;
 
-  @ViewChild('input') private input!: ElementRef<HTMLInputElement>;
+  @ViewChild('errorPanel') protected errorPanel!: OverlayPanel;
 
-  private value: string = '';
+  @ViewChild('input') protected input!: ElementRef<HTMLInputElement>;
 
-  private onValueChange: ((value: string) => void) | null = null;
+  protected value: string = '';
 
-  private onTouched: (() => void) | null = null;
+  protected onValueChange: ((value: string) => void) | null = null;
+
+  protected onTouched: (() => void) | null = null;
 
   public constructor(public control: NgControl) {
     this.control.valueAccessor = this;
@@ -89,10 +77,5 @@ export class TextInputComponent implements DoCheck, ControlValueAccessor {
 
   public setDisabledState?(isDisabled: boolean): void {
     this.disabled = isDisabled;
-  }
-
-  @HostListener('focusout')
-  public onBlur(): void {
-    if (this.onTouched) this.onTouched();
   }
 }
